Added GameScoreObserver.cs; Added IGameModeSimulation interface and GameProgressObserver implementeds it; Now save the level progress as a floating value in the SavingService.

+++ Assets/Code/Scripts/Infrastructure/Services/SavingService/ISavingService.cs
@@ -4,8 +4,8 @@
     {
         void SaveToDisk();
 
-        void SaveLevelProgress(int levelId, int progress);
-        int GetLevelProgress(int levelId);
+        void SaveLevelProgress(int levelId, float progress);
+        float GetLevelProgress(int levelId);
 
         void SaveLongModeLevelProgress(int numberOfScore);
         int GetLongModeProgress();
